31st Commit - Views and Front-End

+++ db/sequel_migrate/40002200_create_tax_bracket.rb
@@ -2,10 +2,10 @@ Sequel.migration do
   up do
     create_table(:tax_bracket) do   
       primary_key :tax_bracket_id
-      BigDecimal :tax_sum
-      BigDecimal :tax_percentage
-      BigDecimal :tax_floor
-      BigDecimal :tax_ceiling
+      Float :tax_sum
+      Float :tax_percentage
+      Float :tax_floor
+      Float :tax_ceiling
       String     :credit_deduction_percentage, :size=>20, :null => false
